apps/api/pyproject.toml
[project]
name = "sidekick-api"
version = "0.1.0"
description = "Sidekick POC API + CLI entry point"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
    "psycopg[binary,pool]>=3.2",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "redis>=5.2",
    "pyyaml>=6.0",
    "typer>=0.15",
    "rich>=13.9",
    "python-dotenv>=1.0",
    "langfuse>=2.50,<3.0",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "qdrant-client>=1.12",
    "sidekick-agent",
    "sidekick-tools",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-xdist>=3.6",
    "respx>=0.22",
    "ruff>=0.8",
    "mypy>=1.13",
]
eval = [
    "ragas>=0.2",
    "deepeval>=1.5",
    "sentence-transformers>=3.3",
    "pandas>=2.2",
    "scipy>=1.14",
    "sidekick-evals",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../../tests"]
